chore(lint): enable clippy::missing_docs in the ui crate
The ui crate has its own [lints] table (no workspace = true inheritance),
so root's missing_docs = "deny" never applied to ui/src despite CI's
clippy/doc jobs running --workspace. Add the missing doc comments in
main.rs and split ShellState/ShellAction/their Reducible impl into a new
shell_state.rs module to keep main.rs under the 500-line linecheck gate.
